requirements.txt
mcp >= 1.26.0
protobuf >= 6.33.5
requests >= 2.32.3
python-dotenv >= 1.0.1